assignment/project/api/requirements.txt
gunicorn==20.1.0
redis==4.1.4
werkzeug==2.0.3
alembic==1.6.5
psycopg2-binary==2.9.1
amqp==5.1.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
black==22.12.0
boto3==1.21.32
botocore==1.24.46
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.1.0
Faker==15.3.4
fastapi==0.88.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
orjson==3.8.3
packaging==22.0
passlib==1.7.4
pathspec==0.10.3
platformdirs==2.6.0
pluggy==1.0.0
prompt-toolkit==3.0.36
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pytest==7.2.0
pytest-asyncio==0.20.3
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
s3transfer==0.5.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.3.24
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
wrapt==1.14.1
opencv-python
torch
numpy
pillow
ultralytics
torchvision
asyncpg